+++ packages/condition_aggregator/lib/condition_aggregator.dart
@@ -45,6 +45,8 @@ export 'src/advisory.dart'
 export 'src/advisory_provider.dart'
     show AdvisoryProvider, AdvisoryProviderInitException;
 export 'src/advisory_absence.dart' show AdvisoryUnavailableReason;
+export 'src/advisory_feed_freshness.dart'
+    show AdvisoryFeedStaleness, AdvisoryFeedFreshnessReporting;
 export 'src/advisory_lookup.dart'
     show
         AdvisoryLookup,